chore: root checking in update script

+++ update.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
 
+if [ "$EUID" -ne 0 ]; then
+  echo "Please run as root"
+  exit
+fi
+
 docker compose down -v
 
 git pull
